requirements.txt
fastapi
uvicorn
pydantic
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
faiss-cpu
pypdf
